Crate asset_test_utils

Module contains predefined test-case scenarios for Runtime with various assets.

Modules

test_cases

Module contains predefined test-case scenarios for Runtime with various assets.

test_cases_over_bridge

Module contains predefined test-case scenarios for Runtime with various assets transferred over a bridge.

xcm_helpers

Helpers for calculating XCM delivery fees.

Macros

include_asset_transactor_transfer_with_local_consensus_currency_works

include_asset_transactor_transfer_with_pallet_assets_instance_works

include_create_and_manage_foreign_assets_for_local_consensus_parachain_assets_works

include_teleports_for_foreign_assets_works

include_teleports_for_native_asset_works

Structs

CollatorSessionKey

CollatorSessionKeys

ExtBuilder

Basic builder based on balances, collators and pallet_session.

RuntimeHelper

SlotDurations

Enums

GovernanceOrigin

Enum representing governance origin/location.

XcmReceivedFrom

Traits

BasicParachainRuntime

A set of traits for a minimal parachain runtime, that may be used in conjunction with the ExtBuilder and the RuntimeHelper.

Functions

assert_metadata

assert_total

mock_open_hrmp_channel

Helper function which emulates opening HRMP channel which is needed for XcmpQueue to pass.

Type Aliases

AccountIdOf

BalanceOf

RuntimeCallOf

RuntimeOriginOf

SessionKeysOf

ValidatorIdOf