asset_test_utils
Function mock_open_hrmp_channel

pub fn mock_open_hrmp_channel<C, T>(
    sender: Id,
    recipient: Id,
    included_head: <<C as Config>::Block as HeaderProvider>::HeaderT,
    slot_digest: &[u8],
)
where
    C: Config,
    T: ProvideInherent<Call = Call<C>>,

Helper function which emulates opening HRMP channel which is needed for XcmpQueue to pass.

Calls parachain-system’s create_inherent in case the channel hasn’t been opened before, and thus requires additional parameters for validating it: latest included parachain head and parachain AuRa-slot.

AuRa consensus hook expects pallets to be initialized, before calling this function make sure to run_to_block at least once.