Function asset_test_utils::test_cases::teleports_for_native_asset_works

pub fn teleports_for_native_asset_works<Runtime, AllPalletsWithoutSystem, XcmConfig, CheckingAccount, WeightToFee, HrmpChannelOpener>(
    collator_session_keys: CollatorSessionKeys<Runtime>,
    slot_durations: SlotDurations,
    existential_deposit: BalanceOf<Runtime>,
    target_account: AccountIdOf<Runtime>,
    unwrap_pallet_xcm_event: Box<dyn Fn(Vec<u8>) -> Option<Event<Runtime>>>,
    runtime_para_id: u32
)
where
    Runtime: Config + Config + Config + Config + Config + Config + Config + Config + Config,
    AllPalletsWithoutSystem: OnInitialize<BlockNumberFor<Runtime>> + OnFinalize<BlockNumberFor<Runtime>>,
    AccountIdOf<Runtime>: Into<[u8; 32]>,
    ValidatorIdOf<Runtime>: From<AccountIdOf<Runtime>>,
    BalanceOf<Runtime>: From<Balance> + Into<u128>,
    WeightToFee: WeightToFee<Balance = Balance>,
    <WeightToFee as WeightToFee>::Balance: From<u128> + Into<u128>,
    <Runtime as Config>::AccountId: Into<<<Runtime as Config>::RuntimeOrigin as OriginTrait>::AccountId> + From<AccountId>,
    <<Runtime as Config>::Lookup as StaticLookup>::Source: From<<Runtime as Config>::AccountId>,
    XcmConfig: Config,
    CheckingAccount: Get<AccountIdOf<Runtime>>,
    HrmpChannelOpener: ProvideInherent<Call = Call<Runtime>>,

Test-case makes sure that Runtime can receive native asset from relay chain and can teleport it back