asset_test_utils

Module test_cases

Module contains predefined test-case scenarios for Runtime with various assets.

Functions

asset_transactor_transfer_with_local_consensus_currency_works

Test-case makes sure that Runtime’s xcm::AssetTransactor can handle native relay chain currency

asset_transactor_transfer_with_pallet_assets_instance_works

Test-case makes sure that Runtime’s xcm::AssetTransactor can handle native relay chain currency

change_storage_constant_by_governance_works

Test-case makes sure that Runtime can change storage constant via governance-like call

create_and_manage_foreign_assets_for_local_consensus_parachain_assets_works

Test-case makes sure that Runtime can create and manage ForeignAssets

reserve_transfer_native_asset_to_non_teleport_para_works

Test-case makes sure that Runtime can reserve-transfer asset to other parachains (where teleport is not trusted)

setup_pool_for_paying_fees_with_foreign_assets

teleports_for_foreign_assets_works

Test-case makes sure that Runtime can receive teleported assets from sibling parachain, and can teleport it back

teleports_for_native_asset_works

Test-case makes sure that Runtime can receive native asset from relay chain and can teleport it back

xcm_payment_api_foreign_asset_pool_works

xcm_payment_api_with_pools_works