Module asset_test_utils::test_cases

Module contains predefined test-case scenarios for Runtime with various assets.

Structs

CollatorSessionKeys

Functions

asset_transactor_transfer_with_local_consensus_currency_works

Test-case makes sure that Runtime’s xcm::AssetTransactor can handle native relay chain currency

asset_transactor_transfer_with_pallet_assets_instance_works

Test-case makes sure that Runtime’s xcm::AssetTransactor can handle native relay chain currency

create_and_manage_foreign_assets_for_local_consensus_parachain_assets_works

teleports_for_foreign_assets_works

Test-case makes sure that Runtime can receive teleported assets from sibling parachain relay chain

teleports_for_native_asset_works

Test-case makes sure that Runtime can receive native asset from relay chain and can teleport it back and to the other parachains